final commit of the  final commit transition to the new version

+++ database/migrations/2020_11_03_175143_create_blog_categories_structure.php
@@ -14,8 +14,8 @@ class CreateBlogCategoriesStructure extends Migration
     public function up()
     {
         Schema::create('blog_categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned()->default(1);
+            $table->bigIncrements('id');
+            $table->bigInteger('parent_id')->unsigned()->default(1);
 
             $table->string('slug')->unique();
             $table->string('title');
@@ -33,6 +33,6 @@ class CreateBlogCategoriesStructure extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('blog_categories');    
+        Schema::dropIfExists('blog_categories');
     }
 }
